+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+1.0.5
+========
+Added ability to have fine grained control over the request that get batched by adding an optional canBatchRequest function on the configuration object.
+
+1.0.4
+========
+Fix to header parsing that would truncate a return header if there was a space it's value i.e. Header: Some Value
+
 1.0.3
 ========
 Build update